requirements.txt
asgiref==3.4.1
pycpfcnpj==1.5.1
Django==3.2.9
django-environ==0.8.1
django-environ-2==2.3.0
djangorestframework==3.12.4
django-filter==2.*
drf-yasg==1.*
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
pytest==6.2.5
pytest-django==4.4.0
pytest-cov==2.*
pytest-factoryboy==2.*
